Arreglando errores con Markdown.Editor
