+++ frontend/tests/controllers/ProblemUpdateTest.php
@@ -1918,13 +1918,9 @@ class ProblemUpdateTest extends \OmegaUp\Test\ControllerTestCase {
             ])
         )['smartyProperties'];
 
-        $this->assertArrayHasKey('problemMarkdownPayload', $response);
-        $this->assertArrayHasKey(
-            'statement',
-            $response['problemMarkdownPayload']
-        );
+        $this->assertArrayHasKey('statement', $response['payload']);
 
-        $originalStatement = $response['problemMarkdownPayload']['statement'];
+        $originalStatement = $response['payload']['statement'];
         $newStatement = [
             'language' => $originalStatement['language'],
             'images' => [],
